requirements.txt
pillow
pyyaml
requests
urllib3<2
flask
gunicorn
